+++ train_swin.py
@@ -18,7 +18,7 @@ from torch.cuda.amp.grad_scaler import GradScaler
 
 from utils.opt import get_opts
 from utils.train_helper import *
-from models.swin import SwinTransformer
+from models.seq_pred import SeqPredictor
 from utils.dataset import CustomDataSet
 from utils.utils import get_summary_writer, save_model
 from utils.cosine_anneal import LECosineAnnealingSmoothRestart
@@ -55,7 +55,7 @@ def setup(args):
         print("CUDA not available.")
         exit(-1)
     
-    swin_model = SwinTransformer(atcg_len, args, emb_dim = 128)
+    swin_model = SeqPredictor(args, emb_dim = 128)
     if not load_path:
         if is_eval:
             raise("LoadPathEmptyError: args.load_path is required in eval mode but not provided.")
@@ -80,9 +80,9 @@ def setup(args):
     if is_eval:
         swin_model.eval()
     else:
-        trainset = None if is_eval else CustomDataSet("./data/", atcg_len, transform, True, args.half_opt)
+        trainset = None if is_eval else CustomDataSet("./data/", atcg_len, transform, True, args.half_opt, mix_up = args.mixup)
         lec_sch = LECosineAnnealingSmoothRestart(args)
-        opt = optim.Adam(params = swin_model.parameters(), lr = lec_sch.lr(epoch), betas=(0.9, 0.999))
+        opt = optim.AdamW(params = swin_model.parameters(), lr = lec_sch.lr(epoch), betas=(0.9, 0.999), weight_decay = weight_decay)
         epochs = args.full_epochs + args.cooldown_epochs
         writer = get_summary_writer(epochs, del_dir)
 
@@ -104,7 +104,7 @@ def train(train_kwargs):
     full_epoch  = train_kwargs['full_epoch'] 
     loss_func   = train_kwargs['loss_func']
 
-    model: SwinTransformer\
+    model: SeqPredictor\
                 = train_kwargs['model']   
     lec_sch: LECosineAnnealingSmoothRestart\
                 = train_kwargs['opt_sch']     
@@ -116,6 +116,8 @@ def train(train_kwargs):
     
     loader_len = len(train_loader)
     for ep in tqdm.tqdm(range(epoch, full_epoch)):
+        if ep > args.mixup_epochs:
+            train_loader.dataset.disable_mixup()
         train_full_num = 0
         train_correct_num = 0
         total_loss = 0
@@ -176,7 +178,7 @@ def eval(eval_kwargs, cur_epoch = 0, use_writer = True, resume = False):
     args        = eval_kwargs['args']
     testset     = eval_kwargs['test_set']   
     loss_func   = eval_kwargs['loss_func']
-    model: SwinTransformer\
+    model: SeqPredictor\
                 = eval_kwargs['model']   
 
     test_loader = DataLoader(testset, args.test_batch_size, shuffle = False, num_workers = 2, drop_last = False)
