+++ utils/dataset.py
@@ -9,13 +9,14 @@ import os
 import torch
 import natsort
 import numpy as np
+from random import randint
 from torch.utils import data
         
 class CustomDataSet(data.Dataset):
     """
         root_dir: dataset root directory, should end with "/"
     """
-    def __init__(self, root_dir: str, seq_num: int, transform = None, is_train = True, use_half = False):
+    def __init__(self, root_dir: str, seq_num: int, transform = None, is_train = True, use_half = False, mix_up = 0.0):
         self.is_train = is_train
         self.root_dir = root_dir
         train_prefix = "train" if is_train else "test"
@@ -28,17 +29,17 @@ class CustomDataSet(data.Dataset):
         img_names = filter(lambda x: x.endswith("dat"), os.listdir(self.data_dir))      # file name inside data_dir is the same with those in label_dir
         all_ndarray = [name for name in img_names]
         self.names = natsort.natsorted(all_ndarray)
+        self.data_num = len(self.names)
         
         self.use_half = use_half
         self.float_type = {"np_float": np.float16, "torch_float": torch.float16} if use_half else {"np_float": np.float32, "torch_float": torch.float32}
 
         if self.transform is None:
             self.transform = lambda x: x
-
-    def __len__(self):
-        return len(self.names)
-
-    def __getitem__(self, idx):
+            
+        self.mix_up = mix_up
+        
+    def load_data(self, idx):
         data_loc = os.path.join(self.data_dir, self.names[idx])
         label_loc = os.path.join(self.label_dir, self.names[idx])
 
@@ -49,6 +50,23 @@ class CustomDataSet(data.Dataset):
         raw_label = np.fromfile(label_loc, dtype = np.uint8).astype(self.float_type["np_float"])
         tensor_label = torch.from_numpy(raw_label)
         return tensor_data, tensor_label
+    
+    def disable_mixup(self):
+        self.mix_up = 0.0
+
+    def __len__(self):
+        return len(self.names)
+
+    def __getitem__(self, idx):
+        seq_1, label_1 = self.load_data(idx)
+        if self.mix_up > 1e-4 and self.is_train:
+            new_idx = (idx + randint(1, self.data_num - 1)) % self.data_num     # make sure when mixing up, there will be no original sequences
+            seq_2, label_2 = self.load_data(new_idx)
+            beta = np.random.beta(self.mix_up, self.mix_up)
+            mixed_seq = beta * seq_1 + (1. - beta) * seq_2
+            mixed_label = beta * label_1 + (1. - beta) * label_2
+            return mixed_seq, mixed_label
+        return seq_1, label_1
 
     def __repr__(self):
         return f"CustomDataSet(\n\tdata_dir={self.data_dir},\n" \
